➖💎 Remove solargraph
Your bundle is locked to parallel (1.20.2) from rubygems repository https://rubygems.org/ or installed locally, but that version can no longer be found in that source. That means the author of parallel (1.20.2) has removed it. You'll need to update your bundle to a version other than parallel (1.20.2) that hasn't been removed in order to install.